Create table food on database then save and get on presentation.

+++ features/main/src/main/java/com/adedom/main/domain/models/FoodModel.kt
@@ -1,10 +1,10 @@
 package com.adedom.main.domain.models
 
 data class FoodModel(
-    val foodId: Int?,
+    val foodId: Long?,
     val foodName: String?,
     val alias: String?,
     val image: String?,
     val ratingScoreCount: String?,
-    val categoryId: Int?,
+    val categoryId: Long?,
 )
